#614: Refactor status change logic around parcels activity and reset selected parcel list from Statuses modal
