+++ src/app/parcels/ActionBar/ActionModals/DeleteParcelModal.tsx
@@ -13,6 +13,7 @@ import { ParcelsTableRow } from "../../parcelsTable/types";
 import DeleteButton from "@/components/Buttons/DeleteButton";
 import DeleteConfirmationDialog from "@/components/Modal/DeleteConfirmationDialog";
 import { ConfirmButtons } from "@/components/Buttons/GeneralButtonParts";
+import { saveDbParcelStatus } from "../saveStatus";
 
 interface ContentProps {
     onClose: () => void;
@@ -80,11 +81,12 @@ const DeleteParcelModal: React.FC<ActionModalProps> = (props) => {
     const numberOfParcelsToDelete = props.selectedParcels.length;
 
     const onDeleteParcels = async (): Promise<void> => {
-        const { error } = await props.updateParcelStatuses(props.selectedParcels, "Parcel Deleted");
+        const { error } = await saveDbParcelStatus(props.selectedParcels, "Parcel Deleted");
         if (error) {
             setErrorMessage(getStatusErrorMessageWithLogId(error));
         } else {
             setSuccessMessage(`${numberOfParcelsToDelete > 1 ? "Parcels" : "Parcel"} Deleted`);
+            props.postSuccessCallback();
         }
         setActionCompleted(true);
     };
