+++ src/app/parcels/ActionBar/Statuses.tsx
@@ -1,18 +1,18 @@
 "use client";
 
-import supabase from "@/supabaseClient";
 import React, { useEffect, useState } from "react";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
-import dayjs, { Dayjs } from "dayjs";
+import { Dayjs } from "dayjs";
 import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
 import StatusesModal from "@/app/parcels/ActionBar/StatusesModal";
-import { logErrorReturnLogId } from "@/logger/logger";
-import { sendAuditLog } from "@/server/auditLog";
 import { ParcelStatus } from "@/databaseUtils";
 import { fetchParcelStatuses } from "@/app/parcels/parcelsTable/fetchParcelTableData";
-
-export type StatusType = ParcelStatus[][number];
+import {
+    saveDbParcelStatus,
+    SaveParcelStatusError,
+    StatusType,
+} from "@/app/parcels/ActionBar/saveStatus";
 
 const nonMenuStatuses: StatusType[] = [
     "Packing Date Changed", //Generated when packing date is changed
@@ -25,71 +25,9 @@ const nonMenuStatuses: StatusType[] = [
 
 const userFacingLabelForEmptyStatus = "No Status";
 
-type SaveParcelStatusErrorType = "eventInsertionFailed";
-
-export interface SaveParcelStatusError {
-    type: SaveParcelStatusErrorType;
-    logId: string;
-}
-
-export interface DeleteClientError {
-    type: SaveParcelStatusErrorType;
-    logId: string;
-}
-
-export type SaveParcelStatusResult = { error: null | SaveParcelStatusError };
-
-export const saveParcelStatus = async (
-    parcelIds: string[],
-    statusName: StatusType,
-    statusEventData?: string | null,
-    action?: string,
-    date?: Dayjs
-): Promise<SaveParcelStatusResult> => {
-    const timestamp = (date ?? dayjs()).toISOString();
-    const eventsToInsert = parcelIds
-        .map((parcelId: string) => {
-            return {
-                new_parcel_status: statusName,
-                parcel_id: parcelId,
-                event_data: statusEventData,
-                timestamp,
-            };
-        })
-        .flat();
-
-    const auditLogs = eventsToInsert.map((eventToInsert) => ({
-        action: action ?? "change parcel status",
-        content: { eventToInsert },
-        parcelId: eventToInsert.parcel_id,
-    }));
-
-    const { data, error } = await supabase
-        .from("events")
-        .insert(eventsToInsert)
-        .select("event_id:primary_key, parcel_id");
-
-    if (error || !data) {
-        const logId = await logErrorReturnLogId("Error with insert: Status event", error);
-        auditLogs.forEach(
-            (auditLog) => void sendAuditLog({ ...auditLog, wasSuccess: false, logId })
-        );
-        return { error: { type: "eventInsertionFailed", logId: logId } };
-    }
-
-    auditLogs.forEach((auditLog) =>
-        sendAuditLog({
-            ...auditLog,
-            eventId: data.find((event) => auditLog.parcelId === event.parcel_id)?.event_id,
-            wasSuccess: true,
-        })
-    );
-
-    return { error: null };
-};
-
 interface Props {
     fetchSelectedParcels: () => Promise<ParcelsTableRow[]>;
+    postSuccessCallback: () => void;
     statusAnchorElement: HTMLElement | null;
     setStatusAnchorElement: React.Dispatch<React.SetStateAction<HTMLElement | null>>;
     setModalError: React.Dispatch<React.SetStateAction<string | null>>;
@@ -110,6 +48,7 @@ const userFacingStatusString = (selectedStatus: StatusType | null): string =>
 
 const Statuses: React.FC<Props> = ({
     fetchSelectedParcels,
+    postSuccessCallback,
     statusAnchorElement,
     setStatusAnchorElement,
     setModalError,
@@ -144,10 +83,8 @@ const Statuses: React.FC<Props> = ({
             setServerErrorMessage("Chosen status was not found.");
             return;
         }
-        const { error } = await saveParcelStatus(
-            selectedParcels.map((parcel: ParcelsTableRow) => {
-                return parcel.parcelId;
-            }),
+        const { error } = await saveDbParcelStatus(
+            selectedParcels,
             selectedStatus,
             null,
             undefined,
@@ -158,6 +95,7 @@ const Statuses: React.FC<Props> = ({
         }
         if (!error) {
             setStatusModal(false);
+            postSuccessCallback();
         }
     };
