+++ src/app/parcels/format.ts
@@ -1,4 +1,4 @@
-import { StatusType } from "./ActionBar/Statuses";
+import { StatusType } from "./ActionBar/saveStatus";
 
 export const formatEventName = (newStatus: StatusType, eventData: string | null): string =>
     `${newStatus} ${eventData ? ` (${eventData})` : ""}`;
